Move SSH timeout copy tests
